requirements.txt
slack-bolt==1.21.2
google-generativeai==0.8.3
chromadb==0.6.3
requests==2.32.3
python-dotenv==1.0.1
